Update NSS to NSS_3_59_1_RTM (m-r84)
This is the update that was previously attempted in UXP Issue 1746 which failed with importing /some/ or all pki-things due to a mangled makefile build system. Now that it is NSS-GYP thanks to my status as a build system god, this version works now.

Moonchild, if you read this. Hope you enjoy having the ability to update NSS again. Oh, btw, you're still a furry cunt.