[DOM:Base] Interpret empty margin string as a zero root  margin in the IntersectionObserver
This trims the whitespace from the margin string first before checking if it is empty, after which, it will be interpreted as a zero root margin. Otherwise, continue with previous behavior (parse margin string using the CSS Parser).

Note: For some reason, Firefox and Chrome treat the unitless zero length as invalid, while with this change, we do not change existing behavior in that regard and continue to accept that value.

Reported on, but does not use any code from https://bugzilla.mozilla.org/show_bug.cgi?id=1738791