[REDO] Bug 1222924 - Stop allowing webpages to link to moz-icon:
In UXP Moonchild changed the implications for this security bug so it actually didn't do what Mozilla intended in an attempt to preserve functionality internally. He failed on both counts. This fix denies web access to any moz-icon but allows it to still work on other protocols like file: and about: etc.

We may want to re-visit the second part of our commit sha 6fa154c0adc64bd43775a79b7b508d87a486882b

Regardless, it seems to now perform as it was intended while not breaking stuff internally.