[Apps:Mail] Re-run and disable original Thunderbird locales via the profile ui migration in prep for support of community-contributed language packs when Ascendant fully lands.
Language Pack policy will have to be established so once we take advantage of the fact that we have easy en-US generation without the former MozNonsense so we can support community-supplied locales or our own.

Though I am not entirely optimistic of the long term sustainability of language packs without either a team to produce them or users with true dedication to their contributed locale. But we shall see!