Revert Bug 495277 as well as altering fireEvent to include XPFE's original extended behavior.
The only functional casualty I see will be the hack to change navigator's urlbar binding to use toolkit autocomplete is the loss of the search for terms xulbox at the bottom of the panel under the tree. Obviously every other autocomplete textbox aside from any other's custom bindings extending autocomplete for this change or switching back and forth between xpfe and toolkit ac to matter.